+++ webcomponents/slides/youtube/src/components/slide/deckdeckgo-slide-youtube.tsx
@@ -149,16 +149,12 @@ export class DeckdeckgoSlideYoutube implements DeckdeckgoSlidePlay {
     });
   }
 
-  private initFrameTitle(): Promise<string> {
-    return new Promise<string>((resolve) => {
-      const title: HTMLElement = this.el.querySelector("[slot='title']");
+  private async initFrameTitle() {
+    const title: HTMLElement = this.el.querySelector("[slot='title']");
 
-      if (title) {
-        this.frameTitle = title.innerHTML;
-      }
-
-      resolve();
-    });
+    if (title) {
+      this.frameTitle = title.innerHTML;
+    }
   }
 
   private initSize(): Promise<void> {