+++ CHANGELOG.md
@@ -9,7 +9,7 @@
 
 ### Others
 
-- deck-utils: v4.0.0 ([CHANGELOG](https://github.com/deckgo/deckdeckgo/blob/master/utils/deck/CHANGELOG.md))
+- deck-utils: v4.0.1 ([CHANGELOG](https://github.com/deckgo/deckdeckgo/blob/master/utils/deck/CHANGELOG.md))
 
 <a name="4.6.0"></a>